Move services to the application project
