+++ PollsAppBlazor/Server/Policy/PollEditAuthorizationHandler.cs
@@ -1,40 +1,39 @@
 ﻿using Duende.IdentityServer.Extensions;
 using Microsoft.AspNetCore.Authorization;
-using PollsAppBlazor.Server.Services;
+using PollsAppBlazor.Application.Services.Implementations;
 
-namespace PollsAppBlazor.Server.Policy
+namespace PollsAppBlazor.Server.Policy;
+
+public sealed class PollEditAuthorizationHandler : PollEditAuthorizationHandler<PollEditAuthorizationRequirement>
 {
-	public sealed class PollEditAuthorizationHandler : PollEditAuthorizationHandler<PollEditAuthorizationRequirement>
-	{
-		public PollEditAuthorizationHandler(PollsService pollsService) : base(pollsService) { }
-	}
+    public PollEditAuthorizationHandler(PollsService pollsService) : base(pollsService) { }
+}
 
-	public abstract class PollEditAuthorizationHandler<TRequirement> : EditAuthorizationHandler<TRequirement>
-		where TRequirement : IAuthorizationRequirement
-	{
-		private readonly PollsService _pollsService;
+public abstract class PollEditAuthorizationHandler<TRequirement> : EditAuthorizationHandler<TRequirement>
+    where TRequirement : IAuthorizationRequirement
+{
+    private readonly PollsService _pollsService;
 
-		public PollEditAuthorizationHandler(PollsService pollsService)
-		{
-			_pollsService = pollsService;
-		}
+    public PollEditAuthorizationHandler(PollsService pollsService)
+    {
+        _pollsService = pollsService;
+    }
 
-		protected override async Task HandleRequirementAsync(AuthorizationHandlerContext context,
-			TRequirement requirement)
-		{
-			if (BasicCheck(context, requirement)) return;
+    protected override async Task HandleRequirementAsync(AuthorizationHandlerContext context,
+        TRequirement requirement)
+    {
+        if (BasicCheck(context, requirement)) return;
 
-			// Check whether user is a creator of the Poll
-			int pollId = GetIntIdFromRoute(context, "pollId");
-			string? creatorId = await _pollsService.GetCreatorIdAsync(pollId);
-			if (creatorId == context.User.GetSubjectId())
-			{
-				context.Succeed(requirement);
-			}
-			else
-			{
-				context.Fail();
-			}
-		}
-	}
+        // Check whether user is a creator of the Poll
+        int pollId = GetIntIdFromRoute(context, "pollId");
+        string? creatorId = await _pollsService.GetCreatorIdAsync(pollId);
+        if (creatorId == context.User.GetSubjectId())
+        {
+            context.Succeed(requirement);
+        }
+        else
+        {
+            context.Fail();
+        }
+    }
 }
